Remove function from shared API
